requirements.txt
streamlit>=1.37,<1.45
pandas>=2.2.3
numpy>=2.1.0
openpyxl>=3.1.2
jieba>=0.42.1
pyyaml>=6.0.1
python-dateutil>=2.9.0
python-docx>=1.1.2
docx2txt>=0.8
pdfplumber>=0.11.4
requests>=2.32.3
beautifulsoup4>=4.12.3
websocket-client>=1.8.0
keyboard>=0.13.5
